Introduce grpc-tools and protobuf 3
Our preference is to avoid platform dependencies as much as possible,
but using protocol buffers will always require the protoc compiler.

Google publish a static binary as part of grpc-tools, and this change
adopts that gem and compiler. As a result of this, we must also change
the base protobuf to version 3. For Digivizer, this means that we must
remove 'optional' and 'required' params from the schema.

Another change is that ruby-compiled protobufs are now suffixed with
`_pb.rb`, from `.pb.rb`. This change affects Smith compilation and
loading, so the compiler has been adapted to account for this.

This changeset is very much WIP, so patience is appreciated!